refactor(cli): split event.go below 1000 lines
Move each event cobra command into a sibling file. Keep
command wiring in event.go. No behavior change.

Related: #652
